+++ src/main/java/de/htwberlin/casino/blackjack/application/domain/service/emitStats/UserStats.java
@@ -8,7 +8,7 @@ package de.htwberlin.casino.blackjack.application.domain.service.emitStats;
  * @param totalBet
  * @param netResult
  */
-public record UserStats(int gamesPlayed,
+public record UserStats(Long gamesPlayed,
                         String winRatio, // Format: "win:loss:tie"
                         int totalBet,
                         int netResult) implements Stats {
